+++ model.py
@@ -26,7 +26,12 @@ class VaseParameters:
     open_top: bool = True
     close_bottom: bool = True
 
+    texture_mode: str = "Pas de texture"
+
     texture_type: str = "Pas de texture"
     texture_zoom: str = "Moyen"
+
+    texture_type_2: str = "Pas de texture"
+    texture_zoom_2: str = "Moyen"
 
     profiles: List[Profile] = field(default_factory=list)